+++ druid/base/static_grid.lua
@@ -63,6 +63,17 @@ local function _extend_border(border, pos, size, pivot)
 end
 
 
+--- Component style params.
+-- You can override this component styles params in druid styles table
+-- or create your own style
+-- @table style
+-- @tfield[opt=false] bool IS_DYNAMIC_NODE_POSES If true, always center grid content as grid pivot sets
+function StaticGrid.on_style_change(self, style)
+	self.style = {}
+	self.style.IS_DYNAMIC_NODE_POSES = style.IS_DYNAMIC_NODE_POSES or false
+end
+
+
 --- Component init function
 -- @tparam StaticGrid self
 -- @tparam node parent The gui node parent, where items will be placed
@@ -191,7 +202,7 @@ function StaticGrid.add(self, item, index, shift_policy)
 	self:_update_indexes()
 	self:_update_borders()
 
-	gui.set_position(item, self:get_pos(index))
+	gui.set_position(item, self:get_pos(index) + self:_get_zero_offset())
 
 	self:_update_pos()
 
@@ -385,8 +396,12 @@ end
 -- @tparam bool is_instant If true, node position update instantly, otherwise with set_position_function callback
 -- @local
 function StaticGrid._update_pos(self, is_instant)
+	local zero_offset = self:_get_zero_offset()
+
 	for i, node in pairs(self.nodes) do
 		local pos = self:get_pos(i)
+		pos.x = pos.x + zero_offset.x
+		pos.y = pos.y + zero_offset.y
 
 		if is_instant then
 			gui.set_position(node, pos)
@@ -396,6 +411,25 @@ function StaticGrid._update_pos(self, is_instant)
 	end
 
 	self.on_update_positions:trigger(self:get_context())
+end
+
+
+--- Return elements offset for correct posing nodes. Correct posing at
+-- parent pivot node (0:0) with adjusting of node sizes and anchoring
+-- @function static_grid:_get_zero_offset
+-- @treturn vector3 The offset vector
+-- @local
+function StaticGrid:_get_zero_offset()
+	if not self.style.IS_DYNAMIC_NODE_POSES then
+		return const.VECTOR_ZERO
+	end
+
+	-- zero offset: center pos - border size * anchor
+	return vmath.vector3(
+		-((self.border.x + self.border.z)/2 + (self.border.z - self.border.x) * self.pivot.x),
+		-((self.border.y + self.border.w)/2 + (self.border.y - self.border.w) * self.pivot.y),
+		0
+	)
 end